Configure and implement webpack bundling
